Fix SparseArray.from_indices dtype on 32-bit platforms
Use sanitize_array in from_indices to ensure consistent int64 dtype
inference, matching SparseArray.__init__ behavior. Also fix test_ufunc
to use explicit int64 dtype for the expected array.

+++ pandas/tests/arrays/sparse/test_array.py
@@ -280,7 +280,7 @@ class TestSparseArrayAnalytics:
 
         sparse = SparseArray([1, -1, 2, -2], fill_value=1)
         result = SparseArray._simple_new(
-            np.array([1, 2, 2]),
+            np.array([1, 2, 2], dtype=np.int64),
             sparse.sp_index,
             SparseDtype(np.int64, 1),
         )
